+++ tabbycat/checkins/urls_public.py
@@ -6,4 +6,7 @@ urlpatterns = [
     path('status/people/',
         views.PublicCheckInPeopleStatusView.as_view(),
         name='public-checkin-status'),
+    path('submit/<slug:url_key>/',
+        views.ParticipantCheckinView.as_view(),
+        name='public-checkin-submit')
 ]